+++ src/gui/FinalMsg.java
@@ -1,4 +1,4 @@
-package co.daylam.gui;
+package gui;
 
 import javax.swing.*;
 import java.awt.*;
