+++ src/gui/MainFrm.form
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="co.daylam.gui.FinalMsg">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="gui.MainFrm">
   <grid id="cbd77" binding="contentPane" layout-manager="GridLayoutManager" row-count="2" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="10" left="10" bottom="10" right="10"/>
     <constraints>
-      <xy x="48" y="54" width="604" height="229"/>
+      <xy x="48" y="54" width="474" height="341"/>
     </constraints>
     <properties/>
     <border type="none" title="Mojtaba Moheb"/>
@@ -34,7 +34,7 @@
                   <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
                 </constraints>
                 <properties>
-                  <text value="Browse webgraphviz"/>
+                  <text value="Generate AST"/>
                 </properties>
               </component>
               <component id="5723f" class="javax.swing.JButton" binding="buttonCancel">
@@ -42,14 +42,14 @@
                   <grid row="0" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
                 </constraints>
                 <properties>
-                  <text value="Exit"/>
+                  <text value="Cancel"/>
                 </properties>
               </component>
             </children>
           </grid>
         </children>
       </grid>
-      <grid id="e3588" layout-manager="GridLayoutManager" row-count="3" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
+      <grid id="e3588" layout-manager="GridLayoutManager" row-count="4" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
         <margin top="0" left="0" bottom="0" right="0"/>
         <constraints>
           <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="0" fill="3" indent="0" use-parent-layout="false"/>
@@ -57,47 +57,59 @@
         <properties/>
         <border type="none"/>
         <children>
-          <component id="827b" class="javax.swing.JTextPane" binding="theASTHasBeenTextPane" default-binding="true">
+          <component id="9f490" class="javax.swing.JRadioButton" binding="fromFileButton">
             <constraints>
-              <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="6" hsize-policy="6" anchor="0" fill="3" indent="0" use-parent-layout="false">
-                <minimum-size width="500" height="50"/>
-                <preferred-size width="500" height="50"/>
-                <maximum-size width="500" height="150"/>
+              <grid row="1" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
+            </constraints>
+            <properties>
+              <selected value="true"/>
+              <text value="Generate AST from the code written in input.java"/>
+            </properties>
+          </component>
+          <component id="7da08" class="javax.swing.JLabel">
+            <constraints>
+              <grid row="0" column="0" row-span="1" col-span="2" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false">
+                <maximum-size width="300" height="500"/>
               </grid>
             </constraints>
             <properties>
-              <editable value="false"/>
-              <enabled value="false"/>
-              <text value="The AST has been exported to output.dot. To visualize it, paste it's content in webgraphviz.com"/>
+              <text value="Choose an option and insert your Java code"/>
             </properties>
           </component>
-          <scrollpane id="82eaa">
+          <component id="92ca7" class="javax.swing.JRadioButton" binding="fromTextAreaButton">
             <constraints>
-              <grid row="2" column="0" row-span="1" col-span="1" vsize-policy="7" hsize-policy="7" anchor="0" fill="3" indent="0" use-parent-layout="false"/>
+              <grid row="2" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
+            </constraints>
+            <properties>
+              <text value="Generate AST from the code written in the below text area"/>
+            </properties>
+          </component>
+          <scrollpane id="d14a">
+            <constraints>
+              <grid row="3" column="0" row-span="1" col-span="2" vsize-policy="7" hsize-policy="7" anchor="0" fill="3" indent="0" use-parent-layout="false"/>
             </constraints>
             <properties/>
             <border type="none"/>
             <children>
-              <component id="49d64" class="javax.swing.JTextArea" binding="textArea1" default-binding="true">
+              <component id="c5944" class="javax.swing.JTextArea" binding="textArea1" default-binding="true">
                 <constraints/>
                 <properties>
-                  <editable value="false"/>
-                  <maximumSize width="600" height="500"/>
-                  <minimumSize width="600" height="500"/>
+                  <editable value="true"/>
+                  <maximumSize width="500" height="300"/>
+                  <minimumSize width="500" height="300"/>
+                  <preferredSize width="500" height="300"/>
                 </properties>
               </component>
             </children>
           </scrollpane>
-          <component id="c6117" class="javax.swing.JLabel">
-            <constraints>
-              <grid row="1" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-            </constraints>
-            <properties>
-              <text value="output.dot:"/>
-            </properties>
-          </component>
         </children>
       </grid>
     </children>
   </grid>
+  <buttonGroups>
+    <group name="buttonGroup1">
+      <member id="92ca7"/>
+      <member id="9f490"/>
+    </group>
+  </buttonGroups>
 </form>
